# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AirDnD API
  version: 0.1.0
  description: AirDnD 프론트엔드와 백엔드가 공유하는 API 계약입니다.
servers:
  - url: http://127.0.0.1:8080
    description: 로컬 백엔드 서버
paths:
...
  /api/rooms:
    get:
      tags: [Rooms]
      summary: 숙소 목록 검색 (커서 페이지네이션)
      description: >
        숙소 목록을 한 페이지씩 검색합니다. 텍스트/필터 조건(region, 날짜, 인원, 가격,
        반려동물)에 더해, 지도 화면에서는 현재 보이는 영역(뷰포트)의 경계 좌표를 함께
        전달해 해당 사각형 영역 안의 숙소만 받아올 수 있습니다.


        페이지네이션:
          - 커서 기반입니다. 첫 요청은 `cursor` 없이 보내고, 다음 페이지는 직전 응답의
            `nextCursor` 를 `cursor` 로 전달합니다. `hasNext=false` 면 마지막 페이지입니다.
          - 한 페이지 크기는 `size`(기본 20)입니다.


        지도 연동 동작:
          - `south`/`west`/`north`/`east` 4개 경계 좌표는 **모두 함께** 전달해야 합니다.
            하나라도 빠지면 경계 필터는 무시됩니다.
          - 4개가 모두 있으면 `south <= latitude <= north` 이고
            `west <= longitude <= east` 인 숙소만 반환합니다.
          - 지도 마커 집계는 별도 엔드포인트 `GET /api/rooms/map` 을 사용하세요.


        날짜 가용성 필터:
          - `checkIn`/`checkOut` 은 **둘 다 또는 둘 다 없이** 전달합니다(한쪽만은 거부).
          - 둘 다 있으면 해당 기간에 예약 불가한 숙소는 **결과에서 제외**됩니다(하드 필터).
          - 날짜가 없으면 가용성 필터 없이 활성 숙소를 모두 보여줍니다
            (이 경우 `isAvailable` 은 항상 true 이며 사실상 장식용입니다).
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: checkIn
          in: query
          schema:
            type: string
            format: date
        - name: checkOut
          in: query
          schema:
            type: string
            format: date
        - name: guests
          in: query
          schema:
            type: integer
            minimum: 1
        - name: adults
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 8
        - name: children
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 8
        - name: infants
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 8
        - name: minPrice
          in: query
          schema:
            type: integer
            minimum: 0
        - name: maxPrice
          in: query
          schema:
            type: integer
            minimum: 0
        - name: allowsPets
          in: query
          schema:
            type: boolean
        - name: south
          in: query
          description: >
            지도 뷰포트 남쪽(최소) 위도. south/west/north/east 4개를 함께 보내면
            그 사각형 영역 안의 숙소만 반환합니다. 프론트엔드 `MapBounds.south`와 1:1.
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: west
          in: query
          description: 지도 뷰포트 서쪽(최소) 경도. `MapBounds.west`와 1:1.
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: north
          in: query
          description: 지도 뷰포트 북쪽(최대) 위도. `MapBounds.north`와 1:1.
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: east
          in: query
          description: 지도 뷰포트 동쪽(최대) 경도. `MapBounds.east`와 1:1.
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: cursor
          in: query
          description: >
            다음 페이지 커서(불투명 문자열). 직전 응답의 `nextCursor` 를 그대로 전달합니다.
            미지정 시 첫 페이지를 반환합니다.
          schema:
            type: string
        - name: size
          in: query
          description: 한 페이지에 담을 숙소 수. 미지정 시 20.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 숙소 목록(커서 페이지)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomSummaryPage"
...
  /api/rooms/{roomId}:
    get:
      tags: [Rooms]
      summary: 숙소 상세 조회
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: 숙소 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/reservations:
    get:
      tags: [Reservations]
      summary: 내 예약 목록 조회
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 예약 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    post:
      tags: [Reservations]
      summary: 예약 생성
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        "201":
          description: 생성된 예약(PENDING 홀드)의 id
          content:
            application/json:
              schema:
                type: integer
                format: int64
                description: 생성된 예약 id. 이후 /checkout/{reservationId} 결제 흐름으로 이어진다.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/reservations/{reservationId}:
    get:
      tags: [Reservations]
      summary: 예약 상세 조회
      security:
        - cookieAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 예약 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Reservations]
      summary: 예약 취소
      security:
        - cookieAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 취소 완료
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/payments/orders:
    post:
      tags: [Payments]
      summary: PayPal 주문 생성
      description: >
        PENDING 예약(reservationId)에 대해 PayPal Orders v2 주문을 만든다.
        결제 금액은 클라이언트가 보내지 않고 서버가 해당 예약에서 재계산한다.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentOrderRequest"
      responses:
        "200":
          description: 생성된 PayPal 주문
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/payments/orders/{orderId}/capture:
    post:
      tags: [Payments]
      summary: 결제 확정(capture)
      description: PayPal 주문을 capture 하고 대상 예약을 PENDING → CONFIRMED 로 확정한다.
      security:
        - cookieAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: PayPal Orders v2 order id
          schema:
            type: string
      responses:
        "200":
          description: 확정된 예약 id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/host/rooms:
    get:
      tags: [Host]
      summary: 호스트 숙소 목록 조회
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 호스트 숙소 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Host]
      summary: 호스트 숙소 등록
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostRoomRequest"
      responses:
        "201":
          description: 등록된 숙소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/host/rooms/{roomId}:
    get:
      tags: [Host]
      summary: 호스트 숙소 상세 조회
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: 호스트 숙소 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Host]
      summary: 호스트 숙소 수정
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostRoomRequest"
      responses:
        "200":
          description: 수정된 숙소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/host/rooms/{roomId}/status:
    patch:
      tags: [Host]
      summary: 호스트 숙소 상태 변경
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [ACTIVE, INACTIVE, PENDING_APPROVAL]
      responses:
        "200":
          description: 상태 변경된 숙소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: 요청 형식 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthenticated:
      description: 로그인 필요
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 비즈니스 충돌
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    RoomSummary:
      type: object
      required:
        [id, name, region, address, pricePerNight, rating, reviewCount, maxGuests, imageUrl, isAvailable, allowsPets]
      properties:
        id:
          type: integer
        name:
          type: string
        region:
          type: string
        address:
          type: string
        pricePerNight:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        maxGuests:
          type: integer
        imageUrl:
          type: string
          format: uri
        isAvailable:
          type: boolean
        allowsPets:
          type: boolean
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: >
            지도 마커 표시용 위도. 지도 검색을 지원하려면 목록 응답에도 포함해야 합니다.
            (백엔드 RoomResponse가 추가하기 전까지는 누락될 수 있어 required 아님)
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: 지도 마커 표시용 경도. latitude와 함께 제공되어야 합니다.
    RoomDetail:
      allOf:
        - $ref: "#/components/schemas/RoomSummary"
        - type: object
          required: [description, amenities, hostName, latitude, longitude]
          properties:
            description:
              type: string
            amenities:
              type: array
              items:
                type: string
            imageUrls:
              type: array
              items:
                type: string
                format: uri
            hostName:
              type: string
            latitude:
              type: number
            longitude:
              type: number
    RoomSummaryPage:
      type: object
      description: 커서 기반 한 페이지. items 가 비고 hasNext=false 면 결과 끝입니다.
      required: [items, nextCursor, hasNext]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RoomSummary"
        nextCursor:
          type: string
          nullable: true
          description: 다음 페이지 커서. hasNext=false 면 null.
        hasNext:
          type: boolean
...
    Reservation:
      type: object
      required: [id, roomId, roomName, roomUrl, checkIn, checkOut, guests, pricePerNight, totalPrice, status]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        roomName:
          type: string
        roomUrl:
          type: string
          format: uri
          description: 숙소 대표 이미지 URL
        region:
          type: string
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        guests:
          type: integer
        pricePerNight:
          type: integer
        totalPrice:
          type: integer
          format: int64
          description: KRW 총액. 서버가 pricePerNight × 박수로 계산하며 클라이언트 입력을 신뢰하지 않는다.
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED]
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: >
            PENDING 홀드 만료 시각. 이 시각 이후의 PENDING 은 점유로 보지 않는다(결제 미완료 자동 해제).
            CONFIRMED/CANCELLED 예약에서는 null.
        guestName:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateReservationRequest:
      type: object
      required: [roomId, checkInDate, checkOutDate, adultCount]
      description: >
        예약(PENDING 홀드) 생성 요청. totalPrice 는 받지 않는다 —
        서버가 숙소 가격과 박수로 계산한다(가격 위변조 방지). guestId 는 인증 세션에서 가져온다.
      properties:
        roomId:
          type: integer
        checkInDate:
          type: string
          format: date
        checkOutDate:
          type: string
          format: date
        adultCount:
          type: integer
          minimum: 1
        childCount:
          type: integer
          minimum: 0
          default: 0
        infantCount:
          type: integer
          minimum: 0
          default: 0
        hasPets:
          type: boolean
          default: false
    CreatePaymentOrderRequest:
      type: object
      required: [reservationId]
      properties:
        reservationId:
          type: integer
          format: int64
          description: 결제 대상 PENDING 예약 id
    CreateOrderResponse:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: string
          description: PayPal Orders v2 order id (capture 시 사용)
    CaptureResponse:
      type: object
      required: [reservationId]
      properties:
        reservationId:
          type: integer
          format: int64
          description: 결제로 확정(CONFIRMED)된 예약 id
    HostRoom:
      allOf:
        - $ref: "#/components/schemas/RoomDetail"
        - type: object
          required: [countryCode, status]
          properties:
            countryCode:
              type: string
              pattern: "^[A-Z]{2}$"
            status:
              type: string
              enum: [ACTIVE, INACTIVE, PENDING_APPROVAL]
    HostRoomRequest:
      type: object
      required:
        [name, region, address, pricePerNight, maxGuests, imageUrl, allowsInfants, allowsPets, countryCode, latitude, longitude]
      properties:
        name:
          type: string
        region:
          type: string
        address:
          type: string
        description:
          type: string
        pricePerNight:
          type: integer
          minimum: 1
        maxGuests:
          type: integer
          minimum: 1
        imageUrl:
          type: string
          format: uri
        imageUrls:
          type: array
          items:
            type: string
            format: uri
        allowsPets:
          type: boolean
          default: false
        allowsInfants:
          type: boolean
          default: false
        countryCode:
          type: string
          pattern: "^[A-Z]{2}$"
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        amenities:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: object
          additionalProperties: true
...